[UI] workaround imgui by drawing more than once (16 though :D)

+++ src/app/view.h
@@ -140,7 +140,7 @@ private:
   double                _perc;
   double                _lastPerc;
   unsigned              _pixels[2];
-  short                 _buffered;
+  int                   _buffered;
   int                   _fixed;
   ViewTabUI*            _tab;
   Window*               _window;
